nixos/pizza: Um, use fileSystems for NFS again
I can no longer reproduce the issue I was experiencing here that waws
mentioned in the comments. I was getting some issues with NFS that I
think were completely unrelated to this but whatever this may or may not
fix it, i.e. I don't have the issues any more.